Conflicts:
	drivers/hid/wacom_wac.c

+++ drivers/gpu/drm/nouveau/nvkm/subdev/bus/hwsq.c
@@ -30,7 +30,7 @@ struct nvkm_hwsq {
 	u32 data;
 	struct {
 		u8 data[512];
-		u8 size;
+		u16 size;
 	} c;
 };
